refactor: move stats tests into pipeline node files with shared test_helpers
Move each operator's stats integration tests to a #[cfg(test)] module in
the operator's own file (sort.rs, filter.rs, scan_source.rs), replacing
the monolithic statistics/tests.rs. Extract shared helpers (data builders,
pipeline construction, end-to-end harness) into pipeline_node::test_helpers.

This eliminates the #[cfg(test)] re-exports of FilterNode, InMemorySourceNode,
ScanSourceNode, and SortNode from pipeline_node/mod.rs — each test module
can now access its node type directly via super::*.
