+++ src/daft-distributed/src/statistics/mod.rs
@@ -1,6 +1,4 @@
 pub(crate) mod stats;
-#[cfg(test)]
-mod tests;
 
 use std::{
     collections::HashMap,
